+++ src/js.rs
@@ -10,6 +10,6 @@ pub fn debugger() {
     crate::api::BabylonApi::debugger();
 }
 
-pub fn random() -> f32 {
+pub fn random() -> f64 {
     crate::api::BabylonApi::random()
 }
