+++ src/math.rs
@@ -1,5 +1,5 @@
 pub use nalgebra::Rotation3;
 pub use nalgebra::Vector3;
 
-pub type Color = Vector3<f32>;
-pub type Vector = Vector3<f32>;
+pub type Color = Vector3<f64>;
+pub type Vector = Vector3<f64>;
